Renamed module from GiftAvailabilityIndicator to GiftAlertIcon
This commit includes renaming files, adjusting namespaces, and identifiers to reflect the module name change. It will not impact the functionality but provides a more accurate description of the module's purpose. This change will make the codebase clearer and easier to maintain in the future.
