+++ LobotomyCorporationMods.GiftAlertIcon/Constants/LocalizationIds.cs
@@ -1,0 +1,12 @@
+﻿// SPDX-License-Identifier: MIT
+
+namespace LobotomyCorporationMods.GiftAlertIcon.Constants
+{
+    internal static class LocalizationIds
+    {
+        internal const string NewGiftTooltip1 = "GiftAlertIcon_NewGiftTooltip1";
+        internal const string NewGiftTooltip2 = "GiftAlertIcon_NewGiftTooltip2";
+        internal const string ReplacementGiftTooltip1 = "GiftAlertIcon_ReplacementGiftTooltip1";
+        internal const string ReplacementGiftTooltip2 = "GiftAlertIcon_ReplacementGiftTooltip2";
+    }
+}
