+++ LobotomyCorporationMods.GiftAlertIcon/Info/GlobalInfo.xml
@@ -1,4 +1,4 @@
 ﻿<?xml version="1.0" encoding="UTF-8"?>
 <info>
-    <ID>GiftAvailabilityIndicator</ID>
+    <ID>GiftAlertIcon</ID>
 </info>
