--- LobotomyCorporationMods.GiftAvailabilityIndicator/Constants/LocalizationIds.cs
@@ -1,12 +1,0 @@
-﻿// SPDX-License-Identifier: MIT
-
-namespace LobotomyCorporationMods.GiftAvailabilityIndicator.Constants
-{
-    internal static class LocalizationIds
-    {
-        internal const string NewGiftTooltip1 = "GiftAvailabilityIndicator_NewGiftTooltip1";
-        internal const string NewGiftTooltip2 = "GiftAvailabilityIndicator_NewGiftTooltip2";
-        internal const string ReplacementGiftTooltip1 = "GiftAvailabilityIndicator_ReplacementGiftTooltip1";
-        internal const string ReplacementGiftTooltip2 = "GiftAvailabilityIndicator_ReplacementGiftTooltip2";
-    }
-}
